+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/mixin/CreativeModeInventoryScreenMixin.java
@@ -1,6 +1,7 @@
 package com.nemonotfound.nemos.inventory.sorting.mixin;
 
 import com.nemonotfound.nemos.inventory.sorting.client.gui.components.AbstractSortButton;
+import com.nemonotfound.nemos.inventory.sorting.factory.DropAllButtonFactory;
 import com.nemonotfound.nemos.inventory.sorting.factory.SortAlphabeticallyButtonFactory;
 import com.nemonotfound.nemos.inventory.sorting.factory.SortAlphabeticallyDescendingButtonFactory;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
@@ -21,16 +22,24 @@ public abstract class CreativeModeInventoryScreenMixin extends AbstractContainer
 
     @Inject(method = "init", at = @At("RETURN"))
     public void init(CallbackInfo ci) {
+        int startIndex = 9;
+        int endIndex = 36;
+        int xOffsetFirstButton = 40;
+        int xOffsetSecondButton = 58;
+        int xOffsetThirdButton = 76;
         int yOffset = 39;
-        int size = 12;
+        int size = 11;
 
         SortAlphabeticallyButtonFactory sortAlphabeticallyButtonFactory = SortAlphabeticallyButtonFactory.getInstance();
         SortAlphabeticallyDescendingButtonFactory sortAlphabeticallyDescendingButtonFactory = SortAlphabeticallyDescendingButtonFactory.getInstance();
+        DropAllButtonFactory dropAllButtonFactory = DropAllButtonFactory.getInstance();
 
-        AbstractSortButton sortAlphabeticallyButton = sortAlphabeticallyButtonFactory.createButton(9, 36, leftPos, topPos, 58, yOffset, imageWidth, size, size, this);
-        AbstractSortButton sortAlphabeticallyDescendingButton = sortAlphabeticallyDescendingButtonFactory.createButton(9, 36, leftPos, topPos, 40, yOffset, imageWidth, size, size, this);
+        AbstractSortButton sortAlphabeticallyDescendingButton = sortAlphabeticallyDescendingButtonFactory.createButton(startIndex, endIndex, leftPos, topPos, xOffsetFirstButton, yOffset, imageWidth, size, size, this);
+        AbstractSortButton sortAlphabeticallyButton = sortAlphabeticallyButtonFactory.createButton(startIndex, endIndex, leftPos, topPos, xOffsetSecondButton, yOffset, imageWidth, size, size, this);
+        AbstractSortButton dropAllButton = dropAllButtonFactory.createButton(startIndex, endIndex, leftPos, topPos, xOffsetThirdButton, yOffset, imageWidth, size, size, this);
 
         this.addRenderableWidget(sortAlphabeticallyButton);
         this.addRenderableWidget(sortAlphabeticallyDescendingButton);
+        this.addRenderableWidget(dropAllButton);
     }
 }